+++ src/fec/fec_encoder.cpp
@@ -62,7 +62,7 @@ uint8_t **FecEncoder::Encode(const char *data, size_t len) {
   fec_params_->encoding_symbol_length = max_size_of_packet_;
 
   if (OF_STATUS_OK != of_set_fec_parameters(fec_session_, fec_params_)) {
-    LOG_ERROR("Set FEC params failed for codec_id {}", fec_codec_id_);
+    LOG_ERROR("Set FEC params failed for codec_id {}", (int)fec_codec_id_);
     return nullptr;
   }
